pyproject.toml
[project]
name = "lzxtools"
version = "0.1.0"
description = "App for getting offers from LZX, Otomot and Pepper and send it trough email"
readme = "README.md"
license = { text = "MIT" }
authors = [{ name = "Somebody ToLove", email = "stl@protonmail.com" }]
requires-python = ">=3.11"
dependencies = [
  "babel",
  "beautifulsoup4",
  "feedparser",
  "imagehash",
  "jinja2",
  "pillow",
  "playwright",
  "python-dateutil",
  "python-dotenv",
  "pytz",
  "schedule",
  "yagmail",
  "zstandard",
]

[project.optional-dependencies]
dev = ["matplotlib", "jupyter"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/lzxtools"]
